Added user input error detection
Fixed a bug where the user could enter anything into the override argument and the code would eventually throw an error. The program will search through the first column and compare with the user input. if it is not found, the program will print out an error message and quit the code. Also added comments